backend/requirements.txt
pandas>=2.0
pyarrow>=14.0
numpy>=1.24
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0
flask-cors>=5.0
requests>=2.31
pywinrm>=0.5.0